feat(TRAVIS_CI): update version

+++ ci/upload_package.py
@@ -24,7 +24,11 @@ if __name__ == "__main__":
 
     if released_version != get_version():
         os.chdir("../")
-        shutil.rmtree('dist')
+
+        # Remove distribution directory if exists
+        if os.path.isdir('dist'):
+            shutil.rmtree('dist')
+
         os.system("python setup.py sdist bdist_wheel")
         os.system("twine upload -r investing-algorithm-framework dist/*")
         os.system("twine upload -p pipy_password -u pipy_username dist/*")
